Update dir structure again, rename templated to category
