+++ src/web/backend/lebanese/layers/category/index.ts
@@ -1,0 +1,1 @@
+export {category} from './category';
