+++ src/web/backend/lebanese/layers/category/libraries/promote/verb.ts
@@ -1,5 +1,5 @@
-import type {ApplyMatchAsType} from "/lib/alphabet/alphabet";
-import {promote as lib} from "/lib/alphabet/library/transform";
+import type {ApplyMatchAsType} from "/lib/alphabet";
+import {promote as lib} from "/lib/transform";
 import {templated} from "../../../templated/templated";
 import {letters} from "../../../underlying/letters";
 import {underlying} from "../../../underlying/underlying";
