+++ src/web/backend/lebanese/layers/underlying/letters.ts
@@ -1,4 +1,4 @@
-import type {ApplyMatchAsType} from "/lib/alphabet/alphabet";
+import type {ApplyMatchAsType} from "/lib/alphabet";
 import type {underlying} from "./underlying";
 
 type Letters = {[K in keyof typeof underlying.types]: Record<string, ApplyMatchAsType<typeof underlying.types[K]>>};
