+++ src/web/backend/lebanese/layers/underlying/underlying.ts
@@ -1,4 +1,4 @@
-import {alphabet} from "/lib/alphabet/alphabet";
+import {alphabet} from "/lib/alphabet";
 
 export const underlying = alphabet({
   name: `underlying`,
